Added setting timer on audio after sit-down sequence

+++ Assets/Scripts/MUserSettings.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 
 public static class MUserSettings
 {
-    private static float time; //in seconds.
+    private static int time; //in seconds.
     private static string environ = "ocean";
 
-    public static float getTime() {
+    public static int getTime() {
         return time;
     }
 
-    public static void setTime(float newTime)
+    public static void setTime(int newTime)
     {
         time = newTime;
     }
